requirements.txt
absl-py==2.2.2
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pillow==11.1.0
protobuf==6.30.2
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.21.0
